Renamed P4CredentialsProvider extensions
