+++ src/main/resources/hudson/plugins/perforce/credentials/P4GlobalCredentialsProvider/global.jelly
