+++ src/main/resources/hudson/plugins/perforce/credentials/P4LocalCredentialsProvider/config.jelly
